requirements.txt
ultralytics
torch
torchvision
opencv-python
numpy
pyyaml


